Restructure a11y to separate script and module from core functionality
