+++ packages/a11y/src/index.js
@@ -2,18 +2,37 @@
  * WordPress dependencies
  */
 import domReady from '@wordpress/dom-ready';
-import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import { makeSetupFunction } from './shared/index';
+import addContainer from './script/add-container';
+import addIntroText from './script/add-intro-text';
+
 export { speak } from './shared/index';
 
 /**
  * Create the live regions.
  */
-export const setup = makeSetupFunction( __( 'Notifications' ) );
+export function setup() {
+	const introText = document.getElementById( 'a11y-speak-intro-text' );
+	const containerAssertive = document.getElementById(
+		'a11y-speak-assertive'
+	);
+	const containerPolite = document.getElementById( 'a11y-speak-polite' );
+
+	if ( introText === null ) {
+		addIntroText();
+	}
+
+	if ( containerAssertive === null ) {
+		addContainer( 'assertive' );
+	}
+
+	if ( containerPolite === null ) {
+		addContainer( 'polite' );
+	}
+}
 
 /**
  * Run setup on domReady.
